convert .qm files back to .ts (#108)
In the original OEM screen binary, we are provided only with `.qm` -- compiled translation -- binaries.  Luckily, `lconvert` can turn this back into a `.ts` XML input, which can be sent to CrowdIn for translation.  In this PR, we:

  * convert original `.qm` files back to `.ts`, and normalize the `.ts` files to the same format that Python would like to generate
  * add some tools to convert `hms_texts` and `error_texts` from JSON to `.ts`
  * add some tools to download translations from CrowdIn and merge them with the existing `.ts` files (and keep them normalized-ish)
  * add a tool to generate a base translation for CrowdIn
  * and interpose the DeviceManager object in bbl_screen, to add arbitrary new translation languages as needed.

While we're in here, we also use `qmllint` on all of our `qml` files, to avoid syntax errors that `lupdate` will cry about (and also, that the runtime will cry about!).  `qmllint` is a pretty piss poor linter but it's better than nothing.

---------

Co-authored-by: Joshua Wise <joshua@joshuawise.com>