interpose: do not attempt to inject languages into something that does not look like a map (#155)
On old firmwares -- like 00.00.24.13 -- the language injection process will
try to peek through data structures in a way that is not valid for that
version of bbl_screen's memory layout.  Now, we make sure that there is at
least *some* kind of QMap where we expect there to be one when we're going
to inject languages.  This means that kexec_ui on 00.00.24.13 will not be
translated correctly, but at least we won't crash.