Avoid leaving bad NumberPads around on the DialogStack if the lock screen is interrupted (#157)
Previously, we used this sort of hokey ersatzDialogStack and tried to mirror the real dialogStack's state onto it (because the real dialogStack was hidden beneath the lock screen). This was fine except if another dialog came in the middle, which would result in the NumberPad becoming very upset, and maybe-or-maybe-not popping the right thing off the stack, and leaving the system in an inconsistent state.

I don't know why I didn't do this to begin with, but this time I just shadow the dialogStack so that the NumberPad refers to the one that is currently active (i.e., the one that is above the lock screen). That way the two are not weirdly intertwined.

While I'm in here, I make sure that the NumberPad is closed when the screen relocks, which seemed like a silly oversight as well.

Fixes #126.

* ScreenLock: instead of using the real (hidden) dialogStack and then cloning its state onto an ersatz dialogStack, just shadow that variable directly to the ersatz one

* ScreenLock: when the screen sleeps, kill the passcode dialog if it is open