Display home status with icon tint
The home button now displays the home status of the printer. Red tint = not homed, green tint = homed

I replaced this ZButton with a ZImage and a MouseArea so that I could use the "tintColor" property of ZImage to make this possible without extra resources.

I also re-enabled the warning dialog that prevents axis movement if the printer is not homed.