Add Wireguard kernel and userspace support. (#164)
Adds a Wireguard kernel module, and revs the kernel itself to have the support infrastructure needed for Wireguard. (While we're in there, we enable cfg80211 debugfs in the kernel.) There are a handful of broken Bambu Lab binaries that get upset about this state of affairs through incorrect usage of getifaddrs; we work around this with -- you guessed it! -- more LD_PRELOAD shenanigans.

Fixes #17.

---------

Co-authored-by: bani6809 <dh585675@anime.net>