ConsolePage: fix wrapping (fixes #145) (#179)
When we added translation support to ConsolePage, we got rid of all of the manual newlines in the placeholder text. Unfortunately, placeholder text is not subject to the wrapMode policy. We simulate placeholder text with real text that is gated on outputText.length -- and then modify the rest of the properties to match the placeholder property. (Since the placeholder text is now plaintext, not rich text, we replace the <br>s with \ns.)