speed presets
Print speed gcode function modified so it accepts either a string or an integer argument. Since the interpolated values are not an exact fit for the data, it's best that users have a choice between using preset values or interpolated values

If a string is input that matches one of the 4 default speed profiles (Silent, Normal, Sport, Ludicrous), it returns gcode that uses preset values for speed parameters. If the argument is an integer, it interpolates the parameters.

Also moved these functions to the end of GcodeGenerator.js after the macros.