SpeedAdjust UI
Created a QML file for a speed adjustment UI which is initialized in AdjustPage.qml.

This uses a Dial type to allow speed adjustment from 30 to 180% with a step size of 2. It interpolates the feed rate and acceleration magnitude values to provide highly consistent extrusion profiles across the entire range.

To maintain support for the default speed profiles, the old UI can still be accessed. When toggling between UIs, a setting is saved to remember which UI to load next time.