use tar2sqfs instead of gensquashfs (#219)
gensquashfs had a bunch of misbehavior in the past, including generating duplicates of /opt/python/bin. Instead, we do something truly terrible, involving generating a tar file with what we want, and then packing that. This sucks but the only thing that sucks worse is what we had, I guess.

At some point the tarfile generation stuff can be refactored out, once we have multiple images. But this is fine for now and unblocks what I was actually trying to do with my night.

* crontab: make +x
* iwconfig: make +x
* squashfs repack: switch to tar2sqfs instead of gensquashfs, which generated weird broken images
* cron: use correct DEVICE_SN