start using the new standard for settings key, and migrate image locations to x1plus settings (#228)
We decided it is easier to name things with dotted settings. So we change locktype to lockscreen.mode and passcode to lockscreen.passcode (and fix the migration logic associated...).

We also migrate cfw_lockscreen_image to lockscreen.image, cfw_home_image to homescreen.image, and cfw_print_image to homescreen.image.printing.

As a bonus of this now being Dbus connected and binding connected, you can now change the image without restarting bbl_screen. Try, for instance:

./x1plus-test-settings.py homescreen.image.printing=/opt/scooter_profile.jpg

(after putting a picture of my former cat in /opt), or:

./x1plus-test-settings.py homescreen.image.printing=None

to return to the default. The screen will change without a reboot!