install.py - error handling and logging (issues 226 and 184)
The issues 184 and 226 indicate a 403 response.

I was able to replicate this while running X1Plus in a separate Python script (403 response when using urllib to download updates from Bambu)

Passing a generic header to urllib fixes the 403 error

I've added urllib.error for better error handling and logging to help debug this.