Added sudo for x1p building
This is more of a useless addition, but it's nice to be able to just copy and paste. In the compiling section, there is no sudo for the commands, and if ran by someone, it will throw an error. Pretty much for lazy people (like me), who just wants to copy and paste.