x1plusd: refactor to split DBus logic into its own class (#261)
Soon, we will want to write hopefully many more DBus-connected services -- for instance, we probably will want to connect gpiokeys to x1plusd... and syslog_shim... and, of course, the OTA daemon. Some of these will have interesting RPCs of their own, and some of them will just want to send signals, but either way, it would be nice not to have to open-code the DBus logic each time. We introduce a X1PlusDBusService class that encapsulates this logic, and rewrite SettingsService in terms of this, and while we are at it, hoist a bunch of the logger control logic out.