Add an X1Plus command line tool (#286)
We add a Python API, x1plus.client, that can be used to programmatically interact with the x1plusd and configure it over the DBus. Additionally, we add a more modern CLI, x1plus, that replaces x1plus-test-ota.py and x1plus-test-settings.py, and provides a more friendly interface (notably, somewhat more free of the None footgun that x1plus-test-settings.py had).

Fixes #252.