use netService in installer instead of wpa_supplicant_hook (#313)
There have been some reports that WiFi works on OEM and works in X1Plus, but not in the installer -- e.g., #309.  In modern OEM firmware (and, indeed, in X1Plus!), the network is managed by `/usr/bin/netService`, rather than by `wpa_supplicant_hook`.  And, in fact, `wpa_supplicant_hook` sometimes can use the wrong config file (though, oddly, in Firmware R, `netService` does not use the "wpa_supplicant_new.conf" like it does in the X1Plus base firmware, but all the same, the user in #309 reports that WiFi works booted into OEM, so ...).

We resolve this in two ways:
* From now on, for all new `kexec_ui`s, we start `netService` instead of `wpa_supplicant_hook` for the early-start WiFi initialization.
* For `install.py` instances running on old `kexec_ui`s, if we detect that `wpa_supplicant_hook` is running, we kill it and restart into `netService`.

While we're in there, we add a little bit of version logging around `install.py`; there are a lot of builds floating around to try to debug this issue, and I realized that we have no way of knowing what build a given install log came from.

Fixes #309.  Hopefully.