OTA and updater: do not attempt to download updates in Shield Mode (#315)
There was, at least, some confusion associated with OTA updates in Shield Mode.  Add some language to make it clearer that both X1Plus OTA and microcontroller firmware update downloads are not supported in Shield Mode.

The `x1plus ota` will still time out attempting to check for updates if the printer is in Shield Mode.  If you are ssh'ed into the printer, you should know what's coming to you.