x1plusd: add a httpd with a simple websocket-to-DBus interface (#337)
We use the `aiohttp` library to add a simple HTTP server that serves some static files (for now, none are actually provided) as well as a WebSocket that provides a JSONRPC interface.  The WebSocket authenticates with a similar mechanism to the VNC password -- i.e., it either grabs a password from the configuration, or it uses the access code.  For now, we provide only one JSONRPC method (after authentication), `dbus.call`, which synchronously (to the websocket) calls an arbitrary DBus method using the Bambu calling convention.

Any exception in the WebSocket is handled gracelessly (i.e., it is logged to syslog, and the connection is instantly dropped).

A sample WebSocket client is shown [here](https://github.com/jwise/x1plusweb/); you can use it online [here](https://jwise.github.io/x1plusweb/) to connect to your printer, but note that you will have to [enable insecure websockets as appropriate for your browser](https://www.damirscorner.com/blog/posts/20210528-allowinginsecurewebsocketconnections.html) when using the GitHub Pages-hosted version.

Future work is to add a mechanism to glue DBus signals and MQTT events into the WebSocket, and to pull in the latest build (or a tagged build, anyway) of x1plusweb when building a .x1p.