Use device_gate to reroute x1plusd sensors MQTT messages to cloud MQTT broker (#353)
It would be nice to get X1Plus MQTT messages into the slicer, not just into the local MQTT broker. Publishing to device/SN/report locally is sufficient to get things into the slicer through the local MQTT broker, but this doesn't translate through the cloud MQTT interface. But device_gate does listen on the local MQTT interface, so we add a new form of device/SN/request, {"x1plus": {"synthesize_report": MSG}} that causes device_gate to rebroadcast MSG on the report topics for all MQTT brokers that it is connected to (cloud and local).

Then, we wire up the x1plusd sensors interface to trigger synthesize_report requests; these, at long last, make it to the cloud, and back to the slicer.