Migrate HardwarePage settings
New X1Plus settings keys:

1. lcd.brightness (float)
backlight brightness

2. leds.toolhead (bool)
toolhead LED on or off

Removed the display brightness timer because we don't need it. The slider's value is saved using an onPressedChanged signal

Also a random thought, but the toolhead LED setting would probably be more useful if it asked users if they want to override the OEM LED state, and if they enable this, then they can choose to have it always on or always off. The way it works right now is .. odd