X1Plus Expander UI (#405)
Adds a handful of UI to drive the Expander, and a small database of Expander configurations. LED strips and LED configurations are still to be done, but this is probably a good start for now. (Realistically, the LED animation system probably needs to change.)

Also left incomplete is having x1plusd actually choose defaults for swapped modules on boot. This PR is really a MVP to get testers up and running with the board, more than anything else.

***

* bbl_screen: add X1Plus Expansion Board logo to Wired Network panel, and correctly render even if the underlying bbl_screen does not support it

* HardwarePage: add UI to show detected Expansion Board, if applicable

* Expansion: UI for Expansion dialog

* Expansion: add an expansion module database, and some scaffolding to load that database, and support for storing metadata in the setting

* Choise: fix the binding loop and text fitment issue

* x1plusd: avoid 3dpsn crash while emulating

* x1plus.js: add Actions to QML, and have button in X1P-005.qml beep the buzzer