Port to 00.00.32.39 (01.08.02.00), and nop out TPU for AMS complaints (#430)
Changes in 00.00.32.39 are relatively minimal; kick everything forward (and make the netservice shim symbolized to be able to port it to other firmware versions in the future). While we're in there, nop out the complaints about TPU from bbl_screen, when it's TPU for AMS, and add a mechanism to be able to specify a custom filament database.

* forward-port netservice_shim to 00.00.32.39
* forward-port bbl_screen UI patches for 00.00.32.39
* update base.json for 01.08.02.00
* add patch to support TPU for AMS
* printer_ui: load filament.zip from x1plusd setting filament.filename
* interpose: fix typo in filament override debug message (thanks, @errge!)